Split the PostProcessingFilter interface in two
