+++ chunky/src/java/se/llbit/chunky/renderer/postprocessing/PostProcessingFilter.java
@@ -18,18 +18,6 @@ public interface PostProcessingFilter {
   void processFrame(int width, int height, double[] input, BitmapImage output, double exposure, TaskTracker.Task task);
 
   /**
-   * Post process a single pixel
-   * @param width The width of the image
-   * @param height The height of the image
-   * @param input The input linear image as double array
-   * @param x The x position of the pixel to process
-   * @param y The y position of the pixel to process
-   * @param exposure The exposure value
-   * @param output The output buffer for the processed pixel
-   */
-  void processPixel(int width, int height, double[] input, int x, int y, double exposure, double[] output);
-
-  /**
    * Get name of the post processing filter
    * @return The name of the post processing filter
    */
