+++ chunky/src/java/se/llbit/chunky/renderer/postprocessing/Tonemap1Filter.java
@@ -6,7 +6,7 @@ import se.llbit.math.QuickMath;
  * Implementation of the tone mapping operator from Jim Hejl and Richard Burgess-Dawson
  * @link http://filmicworlds.com/blog/filmic-tonemapping-operators/
  */
-public class Tonemap1Filter extends PixelPostProcessingFilter {
+public class Tonemap1Filter extends SimplePixelPostProcessingFilter {
   @Override
   public void processPixel(double[] pixel) {
     for(int i = 0; i < 3; ++i) {
